refactor(backend): refactored user registration-related files
