+++ power-pay-backend/src/main/java/com/adorsys/gis/powerpay/powerpaybackend/domain/FirstRegistrationRequest.java
@@ -1,0 +1,12 @@
+package com.adorsys.gis.powerpay.powerpaybackend.domain;
+import lombok.Getter;
+import lombok.Setter;
+
+
+
+@Getter
+@Setter
+public class FirstRegistrationRequest {
+    String phoneNumber;
+    String userName;
+}
