+++ power-pay-backend/src/main/java/com/adorsys/gis/powerpay/powerpaybackend/domain/SecondRegistrationRequest.java
@@ -1,0 +1,13 @@
+package com.adorsys.gis.powerpay.powerpaybackend.domain;
+import lombok.Getter;
+import lombok.Setter;
+
+
+
+@Getter
+@Setter
+public class SecondRegistrationRequest {
+    String pin;
+    String otp;
+
+}
